+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - This CHANGELOG file.
+- Statute-level percentages to CSP state distribution data JSON. [#12](https://github.com/policy-design-lab/data-import/issues/12)
 
 ### Changed
 
@@ -20,3 +21,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Title 1 Commodities JSON files by adding zero entries. [#6](https://github.com/policy-design-lab/data-import/issues/6)
+- Title 2 CSP JSON files by adding zero entries. [#10](https://github.com/policy-design-lab/data-import/issues/10)